Refactoring in helpers
