--- src/helpers/getRandomBool.js
@@ -1,3 +1,0 @@
-export const getRandomBool = () => {
-  return Math.random() > 0.5;
-};
