--- src/helpers/getSimpleHandler.js
@@ -1,11 +1,0 @@
-export const getSimpleHandler = (selector, handler) => {
-  return (event) => {
-    const control = event.target.closest(selector);
-
-    if (!control) {
-      return;
-    }
-
-    handler();
-  };
-};
