+++ src/helpers/index.js
@@ -1,15 +1,9 @@
-export {getDate, getTime, getPeriod} from './getDateTime';
-export {createElement} from './createElement';
-export {renderElement} from './renderElement';
-export {replaceElement} from './replaceElement';
-export {removeElement} from './removeElement';
-export {getClass} from './getClass';
-export {getRandomBool} from './getRandomBool';
+export {getDate, getTime, getPeriod} from './date-time';
+export {createElement, removeElement, renderElement, replaceElement} from './dom';
+export {getClass} from './get-class';
 export {getTodayTasks, getOverdueTasks, getFilteredTasks} from './filter';
 export {getSortedTasks} from './sort';
-export {getSimpleHandler} from './getSimpleHandler';
-export {getHandlerWithProp} from './getHandlerWithProp';
-export {getHandlerWithValue} from './getHandlerWithValue';
+export {getSimpleHandler, getHandlerWithProp, getHandlerWithValue} from './handlers';
 export {getTasksQuantityByColor, getTasksQuantityByDay, getArchivedTasksBetweenFromTo, getDaysBetweenFromTo} from './statistic';
-export {getPlurals} from './getPlurals';
+export {getPlurals} from './get-plurals';
 export {shake} from './shake';
