--- src/helpers/removeElement.js
@@ -1,4 +1,0 @@
-export const removeElement = (component) => {
-  component.getElement().remove();
-  component.removeElement();
-};
