--- src/helpers/replaceElement.js
@@ -1,7 +1,0 @@
-export const replaceElement = (componentFrom, componentTo) => {
-  if (!componentFrom && !componentTo) {
-    return;
-  }
-
-  componentFrom.getElement().replaceWith(componentTo.getElement());
-};
